requirements.txt
streamlit
playwright
langchain
langgraph
langchain-openai
langchain-community
pymongo
python-dotenv
beautifulsoup4
pandas
lxml
